requirements.txt
flask
flask-sqlalchemy
python-dotenv
flask-socketio
eventlet
websocket-client
requests
